refactor : View에 interface를 추가
