+++ src/main/java/carracing/CarRacingApplication.java
@@ -3,13 +3,13 @@ package carracing;
 import carracing.controller.CarRacingController;
 import carracing.domain.service.RandomFourStepRacingService;
 import carracing.ui.InputView;
-import carracing.ui.RacingView;
+import carracing.ui.RacingViewImpl;
 import carracing.ui.ResultView;
 
 public class CarRacingApplication {
 
   public static void main(String[] args) {
-    RacingView racingView = new RacingView(InputView.getInstance(), ResultView.getInstance());
+    RacingViewImpl racingView = new RacingViewImpl(InputView.getInstance(), ResultView.getInstance());
     RandomFourStepRacingService carRacingService = new RandomFourStepRacingService();
 
     CarRacingController carRacingController = new CarRacingController(carRacingService, racingView);
