+++ src/main/java/carracing/controller/CarRacingController.java
@@ -2,14 +2,14 @@ package carracing.controller;
 
 import carracing.domain.entity.Challengers;
 import carracing.domain.service.CarRacingService;
-import carracing.ui.RacingView;
+import carracing.ui.RacingViewImpl;
 
 public class CarRacingController {
 
   private final CarRacingService carRacingService;
-  private final RacingView racingView;
+  private final RacingViewImpl racingView;
 
-  public CarRacingController(CarRacingService carRacingService, RacingView racingView) {
+  public CarRacingController(CarRacingService carRacingService, RacingViewImpl racingView) {
     this.carRacingService = carRacingService;
     this.racingView = racingView;
   }
